frameSecurity/requirements.txt
opencv-python
numpy
pycryptodome
tqdm
ultralytics
pillow
gradio